Add getMethodCalls, tests for captured args
Add the getMethodCalls function (_mightSpy_getMethodCalls(<methodName>)
in the returned proxy object).  Returns an array of caputred method
calls, or an empty array if nothing has been captured for the given
method.  Does not throw an error, as we don't know if the proxied object
has an onMissingMethod handler.  This is epsecially important when
spying a mock.