Added pause/record capability
When creating the proxy, init accepts an optional boolean argument.  If
false, spy does not record any method calls.  Defaults to true.

Also added 'pause' (_mightyMock_pause in the proxied object) and
'record) (_mightyMock_record) to stop/resume recording of method calls.

This is useful when spying a mock, as the initial mock method setup can
be ignored, and record mode turned on once mock is ready to be used.